% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpayers_gal.R
\name{allpayers_gal}
\alias{allpayers_gal}
\title{allpayers_gal: clean Galileo file for binding with result of allpayers_mco function}
\usage{
allpayers_gal(g, c)
}
\value{
Tibble

g = Galileo 'Capitation List' file
c = CurrentHIEIDs file
}
\description{
allpayers_gal: clean Galileo file for binding with result of allpayers_mco function
}
\examples{
allpayers_gal(g, c)

}
