% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpayers_mco.R
\name{allpayers_mco}
\alias{allpayers_mco}
\title{allpayers_mco: clean AllPayersHIEID file for binding with result of allpayers_galileo function}
\usage{
allpayers_mco(a, c)
}
\value{
Tibble

a = AllPayersHIEID file
c = CurrentHIEIDs file
}
\description{
allpayers_mco: clean AllPayersHIEID file for binding with result of allpayers_galileo function
}
\examples{
allpayers_mco(a, c)

}
