% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camcare.R
\name{camcare}
\alias{camcare}
\title{CAMcare Function: clean CAMcare patient list}
\usage{
camcare(x)
}
\value{
Tibble
}
\description{
CAMcare Function: clean CAMcare patient list
}
\examples{
camcare(x)
y <- camcare(x) \%>\% write_csv(y, "y.csv", na = "")
}
