% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate}
\alias{deduplicate}
\title{allpayers_mco: clean AllPayersHIEID file for binding with result of allpayers_galileo function}
\usage{
deduplicate(m, g)
}
\value{
Tibble

a = AllPayersHIEID file
c = CurrentHIEIDs file
}
\description{
allpayers_mco: clean AllPayersHIEID file for binding with result of allpayers_galileo function
}
\examples{
allpayers_mco(a, c)

}
