% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galileo.R
\name{galileo}
\alias{galileo}
\title{Galileo Function: clean Galileo capitation list}
\usage{
galileo(x)
}
\value{
Tibble
}
\description{
Galileo Function: clean Galileo capitation list
}
\examples{
galileo(x)
y <- galileo(x) \%>\% write_csv(y, "y.csv", na = "")
}
