% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united.R
\name{united}
\alias{united}
\title{United Function: clean United patient list}
\usage{
united(x)
}
\value{
Tibble
}
\description{
United Function: clean United patient list
}
\examples{
united(x)
y <- united(x) \%>\% write_csv(y, "y.csv", na = "")
}
